refactor(cpp): isolate LoaderPluginConfig
